create package for pages
