--- src/test/java/pageobjects/BasePage.java
@@ -1,11 +1,0 @@
-package pageobjects;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-public class BasePage {
-
-
-
-}
